Make APM disk I/O portable and large-file safe
Replace packed disk-structure overlays with explicit big-endian serialization and safe driver-map accessors.

Use 64-bit file offsets and checked block arithmetic so 32-bit builds can create partitions above 2 GiB while enforcing the Apple Partition Map address-space limit.

Add wire-format and large-offset regression tests.

+++ partition_map.h
@@ -79,7 +79,7 @@ partition_map* find_entry_by_sector(uint32_t lba, partition_map_header *map);
 partition_map_header* init_partition_map(char *name, partition_map_header* oldmap);
 void move_entry_in_map(long old_index, long index, partition_map_header *map);
 partition_map_header* open_partition_map(char *name, int *valid_file);
-void resize_map(long new_size, partition_map_header *map);
+void resize_map(uint32_t new_size, partition_map_header *map);
 void write_partition_map(partition_map_header *map);
 uint32_t find_free_space(partition_map_header *map);
